[cleanup] Moved all custom helper scripts to scripts dir
