+++ install/debian/install.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env sh
 
 # bring in the variables: `normal`, `info`, `warning`, `danger`, `reset`, `user_agent`
-. "$(dirname "$(readlink -f "$0")")/../utils.sh"
+. "$(dirname "$(readlink -f "$0")")/../utils/utils.sh"
 
 IsInstalled() {
   directory=$1
@@ -16,7 +16,7 @@ RootDir=$1
 
 ########### Pip is the foremost thing that must be installed along with some needed dependencies for python libraries
 
-apt_wrapper_path="$RootDir/install/aptitude-wrapper.sh"
+apt_wrapper_path="$RootDir/install/utils/aptitude-wrapper.sh"
 
 # Perform apt-get update before starting to install all packages, so we can get the latests manifests and packages versions
 sudo apt-get update
