+++ install/distro-independent.cfg
@@ -33,17 +33,17 @@ command = wget --user-agent="Mozilla/5.0 (X11; Linux i686; rv:6.0) Gecko/2010010
 # This directory cannot exist, so the script will always run. Checks for individual ca.crt and folders is present inside
 # the script so, it won't be a problem.
 directory = /tmp/owtf-install/ca/%(Pid)s
-command = sh %(RootDir)s/install/proxy_CA.sh %(RootDir)s
+command = sh %(RootDir)s/scripts/proxy_CA.sh %(RootDir)s
 
 [Database Config]
 # This directory cannot exist, so the script will always run. Checks for individual db.cfg is present inside
 # the script so, it won't be a problem.
 directory = /tmp/owtf-install/db-config/%(Pid)s
-command = sh %(RootDir)s/install/db_config_setup.sh %(RootDir)s
+command = sh %(RootDir)s/scripts/db_config_setup.sh %(RootDir)s
 
 [Zest Jars]
 directory = /tmp/owtf-install/zest/%(Pid)s
-command = sh %(RootDir)s/install/install_zest_jars.sh
+command = sh %(RootDir)s/scripts/install_zest_jars.sh
 
 [CMS Explorer]
 directory = %(RootDir)s/tools/restricted/cms-explorer
