+++ scripts/update_convert_cms_explorer_dicts.sh
@@ -3,7 +3,7 @@
 
 
 # bring in the variables: `normal`, `info`, `warning`, `danger`, `reset`, `user-agent`
-. "$(dirname "$(readlink -f "$0")")/utils.sh"
+. "$(dirname "$(readlink -f "$0")")/../install/utils/utils.sh"
 
 RootDir=$1
 SOURCE_DIR="$RootDir/tools/restricted/cms-explorer"
